+++ src/any_vec_typed.rs
@@ -5,7 +5,7 @@ use std::ptr::NonNull;
 use std::slice;
 use crate::any_value::{AnyValue, AnyValueWrapper};
 use crate::any_vec_raw::AnyVecRaw;
-use crate::ops::{ElementIter, remove, swap_remove, TempValue};
+use crate::ops::{Iter, remove, swap_remove, TempValue};
 use crate::any_vec_ptr::AnyVecRawPtr;
 use crate::into_range;
 use crate::iter::ElementIterator;
@@ -90,7 +90,7 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
         -> impl ElementIterator<Item = T>
     {
         let Range{start, end} = into_range(self.len(), range);
-        ElementIter(Drain::new(
+        Iter(Drain::new(
             AnyVecRawPtr::<T>::from(self.any_vec),
             start,
             end
@@ -110,7 +110,7 @@ impl<'a, T: 'static> AnyVecTyped<'a, T>{
         let replace_with = replace_with.into_iter()
             .map(|e| AnyValueWrapper::new(e));
 
-        ElementIter(Splice::new(
+        Iter(Splice::new(
             AnyVecRawPtr::<T>::from(self.any_vec),
             start,
             end,