+++ src/ops/mod.rs
@@ -1,12 +1,12 @@
 mod temp;
-mod element_iter;
+mod iter;
 pub(crate) mod swap_remove;
 pub(crate) mod remove;
 pub(crate) mod drain;
 pub(crate) mod splice;
 
 pub use temp::TempValue;
-pub use element_iter::ElementIter;
+pub use iter::Iter;
 
 use crate::any_vec_ptr::AnyVecPtr;
 
@@ -24,12 +24,22 @@ pub type Remove<'a, Traits> = TempValue<remove::Remove<'a, AnyVecPtr<Traits>>, T
 /// [`AnyVec::swap_remove`]: crate::AnyVec::swap_remove
 pub type SwapRemove<'a, Traits> = TempValue<swap_remove::SwapRemove<'a, AnyVecPtr<Traits>>, Traits>;
 
-/// Same as [`std::vec::Drain`].
+///  A draining [`ElementIterator`] for [`AnyVec`]. Return [`Element`] items.
 ///
 /// This is created by [`AnyVec::drain`].
 ///
+/// [`AnyVec`]: crate::AnyVec
 /// [`AnyVec::drain`]: crate::AnyVec::drain
-pub type Drain<'a, Traits> = ElementIter<drain::Drain<'a, AnyVecPtr<Traits>>>;
+/// [`Element`]: crate::element::Element
+/// [`ElementIterator`]: crate::iter::ElementIterator
+pub type Drain<'a, Traits> = Iter<drain::Drain<'a, AnyVecPtr<Traits>>>;
 
-// TODO: description!!!
-pub type Splice<'a, Traits, I> = ElementIter<splice::Splice<'a, AnyVecPtr<Traits>, I>>;
+///  A splicing [`ElementIterator`] for [`AnyVec`]. Return [`Element`] items.
+///
+/// This is created by [`AnyVec::splice`].
+///
+/// [`AnyVec`]: crate::AnyVec
+/// [`AnyVec::splice`]: crate::AnyVec::splice
+/// [`Element`]: crate::element::Element
+/// [`ElementIterator`]: crate::iter::ElementIterator
+pub type Splice<'a, Traits, I> = Iter<splice::Splice<'a, AnyVecPtr<Traits>, I>>;